requirements.txt
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
python-dateutil==2.8.2
websockets==12.0